Move the singular (domain-forwarding) components and routes to the plural form (domain-forwardings)
